refactor: S3 관련 external.s3로 리팩토링
