+++ src/main/java/com/hanium/mom4u/global/config/WebConfig.java
@@ -1,4 +1,4 @@
-package com.hanium.mom4u.external;
+package com.hanium.mom4u.global.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
